+++ test/test.js
@@ -546,7 +546,7 @@ describe('URDFExporter', () => {
 
         it.skip('should modify the directory of textures included with the exported meshes', () => {});
 
-        it.skip('should include material properties and textures when `includesMaterials` is `false`', () => {});
+        it.skip('should include material properties and textures when `material` is included from the mesh generation function', () => {});
 
     });